brain/rag_requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
banks==2.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
dataclasses-json==0.6.7
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
faiss-cpu==1.12.0
filelock==3.19.1
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.9.0
greenlet==3.2.4
griffe==1.14.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.1
idna==3.10
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.30
langchain-core==0.3.76
langchain-experimental==0.3.4
langchain-huggingface==0.3.1
langchain-text-splitters==0.3.11
langsmith==0.4.31
llama-index-core==0.14.3
llama-index-instrumentation==0.4.1
llama-index-llms-groq==0.4.1
llama-index-llms-openai==0.6.1
llama-index-llms-openai-like==0.5.2
llama-index-workflows==2.5.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numpy==2.3.3
openai==1.109.1
orjson==3.11.3
packaging==25.0
pillow==11.3.0
platformdirs==4.4.0
propcache==0.3.2
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
PyMuPDF==1.26.4
python-dotenv==1.1.1
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
setuptools==80.9.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.56.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
wrapt==1.17.3
yarl==1.20.1
zstandard==0.25.0
